requirements.txt
pytest~=8.3.4
requests~=2.32.3
PyYAML~=6.0.2

selenium~=4.26.1
allure-python-commons~=2.14.3
loguru~=0.7.3
Flask~=3.1.1
h11~=0.14.0
pip~=23.2.1
attrs~=24.2.0
retry~=0.9.2
PySocks~=1.7.1
pillow~=11.0.0
tornado~=6.4.2
Jinja2~=3.1.4
cffi~=1.17.1
setuptools~=75.6.0
typing_extensions~=4.12.2
trio~=0.27.0
outcome~=1.3.0.post0
sniffio~=1.3.1
sortedcontainers~=2.4.0
idna~=3.10
MarkupSafe~=3.0.2
pluggy~=1.5.0
Deprecated~=1.2.15
iniconfig~=2.0.0
py~=1.11.0
numpy~=2.2.1
certifi~=2024.8.30
urllib3~=2.2.3
trio-websocket~=0.11.1
websocket-client~=1.8.0
packaging~=24.2
pycparser~=2.22
wsproto~=1.2.0